infrastructure/requirements.txt
# Dependências para scripts de infraestrutura e automação
fabric>=3.2
python-dotenv>=1.0
motor>=3.3
sqlalchemy[asyncio]>=2.0
asyncpg>=0.27
alembic>=1.12
